src/app/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
psycopg2-binary>=2.9.9
httpx>=0.25.0
pydantic>=2.5.0
requests>=2.31.0
